Student refactored.

+++ backend-ms/university/src/main/java/com/ugc/university/repository/course/StreamRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StreamRepository extends JpaRepository<Stream, Long> {
+public interface StreamRepository extends JpaRepository<Stream, Integer> {
 }
